requirements.txt
asn1crypto==0.24.0
blinker==1.4
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.6.1
Flask==1.0.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.3.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
Pillow==6.0.0
pyaes==1.6.1
pycparser==2.19
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.3
urllib3==1.24.2
Werkzeug==0.15.2
wincertstore==0.2
